requirements.txt
python-telegram-bot==20.7
spleeter==2.3.2